requirements.txt
Django
django-localflavor

